requirement.txt
certifi==2019.3.9
chardet==3.0.4
Django==1.11.20
django-cors-headers==3.0.1
djangorestframework==3.9.4
idna==2.8
mutagen==1.42.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
requests==2.21.0
urllib3==1.24.1
